medium_/requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.6
django-filter==1.1.0
django-phonenumber-field==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.7
idna-ssl==1.0.1
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
multidict==4.3.1
oauthlib==2.1.0
openapi-codec==1.3.2
phonenumberslite==8.9.8
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
telepot==12.7
uritemplate==3.0.0
urllib3==1.23
yarl==1.2.6
